Move tests to sdk-tests, move autoconfig to starter
