+++ dapr-spring/dapr-spring-boot-autoconfigure/src/main/java/io/dapr/spring/boot/autoconfigure/statestore/DaprStateStoreAutoConfiguration.java
@@ -25,7 +25,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.data.keyvalue.core.KeyValueAdapter;
 
 @AutoConfiguration(after = DaprClientAutoConfiguration.class)
@@ -53,7 +52,6 @@ public class DaprStateStoreAutoConfiguration {
   }
 
   @Bean
-  @Lazy
   @ConditionalOnMissingBean
   public DaprKeyValueTemplate keyValueTemplate(KeyValueAdapterResolver keyValueAdapterResolver) {
     return new DaprKeyValueTemplate(keyValueAdapterResolver);
