+++ dapr-spring/dapr-spring-boot-testcontainers/pom.xml
@@ -20,16 +20,6 @@
       <artifactId>spring-test</artifactId>
     </dependency>
     <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-web</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-testcontainers</artifactId>
-        <scope>test</scope>
-    </dependency>
-    <dependency>
       <groupId>org.testcontainers</groupId>
       <artifactId>testcontainers</artifactId>
     </dependency>
@@ -47,12 +37,6 @@
       <groupId>io.dapr</groupId>
       <artifactId>testcontainers-dapr</artifactId>
       <version>${testcontainers-dapr.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>io.dapr</groupId>
-      <artifactId>dapr-sdk-workflows</artifactId>
-      <version>${dapr-workflows.sdk.version}</version>
-      <scope>test</scope>
     </dependency>
   </dependencies>
 </project>
